+++ examples/presentation/common.go
@@ -16,24 +16,24 @@ type WorldInterface interface {
 
 type Square struct {
 	Rect        *resolv.Rectangle
-	SpeedX      float32
-	SpeedY      float32
-	BounceFrame float32
+	SpeedX      float64
+	SpeedY      float64
+	BounceFrame float64
 }
 
 func NewSquare(space *resolv.Space) *Square {
 
-	square := &Square{Rect: resolv.NewRectangle(cell*2+float32(rand.Int31n(screenWidth-cell*4)), cell*2+float32(rand.Int31n(screenHeight-cell*4)), cell, cell),
-		SpeedX: (0.5 - rand.Float32()) * 8,
-		SpeedY: (0.5 - rand.Float32()) * 8}
+	square := &Square{Rect: resolv.NewRectangle(cell*2+float64(rand.Int31n(screenWidth-cell*4)), cell*2+float64(rand.Int31n(screenHeight-cell*4)), cell, cell),
+		SpeedX: (0.5 - rand.Float64()) * 8,
+		SpeedY: (0.5 - rand.Float64()) * 8}
 
 	// Attempt to not spawn a Square in an occupied location
 	for i := 0; i < 100; i++ {
 
 		if space.IsColliding(square.Rect) {
 
-			square.Rect.X = cell*2 + float32(rand.Int31n(screenWidth-cell*4))
-			square.Rect.Y = cell*2 + float32(rand.Int31n(screenHeight-cell*4))
+			square.Rect.X = cell*2 + float64(rand.Int31n(screenWidth-cell*4))
+			square.Rect.Y = cell*2 + float64(rand.Int31n(screenHeight-cell*4))
 
 		}
 
@@ -49,12 +49,12 @@ func NewSquare(space *resolv.Space) *Square {
 
 }
 
-func DrawText(x, y float32, textLines ...string) {
+func DrawText(x, y float64, textLines ...string) {
 
 	for i, line := range textLines {
 
 		// length := rl.MeasureText(line, 8)
-		// rl.DrawRectangle(x-2, y+(float32(i)*10), length+2, 8, rl.Black)
+		// rl.DrawRectangle(x-2, y+(float64(i)*10), length+2, 8, rl.Black)
 		rl.DrawText(line, int32(x), int32(y)+int32(i*10), 8, rl.Blue)
 		rl.DrawText(line, int32(x), int32(y)-1+int32(i*10), 8, rl.White)
